+++ examples/compare_kps3_kps4.jl
@@ -38,7 +38,7 @@ function simulate(s, integrator, steps, plot=false; fig="")
             println("lift, drag  [N]: $(round(lift, digits=2)), $(round(drag, digits=2))")
         end
 
-        KiteModels.next_step!(s, integrator, dt=dt)
+        KiteModels.next_step!(s, integrator; set_speed=0, dt=dt)
         
         if plot
             reltime = i*dt-dt