+++ mwes/mwe_02.jl
@@ -8,11 +8,11 @@ kps4::KPS4 = KPS4(KCU(set))
 STEPS = 200
 
 integrator = KiteModels.init_sim!(kps4; stiffness_factor=0.035, prn=false)
-next_step!(kps4, integrator, v_ro = 0, dt=0.05)
-@timev next_step!(kps4, integrator, v_ro = 0, dt=0.05)
+next_step!(kps4, integrator, set_speed = 0, dt=0.05)
+@timev next_step!(kps4, integrator; set_speed = 0, dt=0.05)
 function nsteps(kps4, integrator)
     for i=1:STEPS
-        next_step!(kps4, integrator, v_ro = 0, dt=0.05)
+        next_step!(kps4, integrator; set_speed = 0, dt=0.05)
     end
 end
 bytes = @allocated nsteps(kps4, integrator)